Move main_window, app_state, default_state under src/ui/
Move src/main_window.py, src/app_state.py, src/default_state.py to
src/ui/ using git mv. Update src/main.py import to use .ui.main_window.
Fix main_window.py imports from .ui.handlers and .ui.widgets to relative
.handlers and .widgets since main_window is now in src/ui/. All other
relative imports remain valid after the moves.
