+++ src/main.py
@@ -34,7 +34,7 @@ import sys
 
 from PyQt5.QtWidgets import QApplication
 
-from .main_window import MainWindow
+from .ui.main_window import MainWindow
 
 
 def main() -> None:
